third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/elu.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] elu float32 4D tensor positive options.alpha
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Setting a custom alpha is not supported in tflite schema."
[FAIL] elu float32 4D tensor negative options.alpha
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Setting a custom alpha is not supported in tflite schema."
[FAIL] elu float16 positive 0D scalar default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test elu float16 actual 0 should be close enough to expected 4.72265625 by ULP distance: expected a number less than or equal to 18 but got 17593
[FAIL] elu float16 negative 0D scalar default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test elu float16 actual 0 should be close enough to expected -0.97900390625 by ULP distance: expected a number less than or equal to 18 but got 48085
[FAIL] elu float16 1D constant tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test elu float16 actual 0.0078125 should be close enough to expected 4.72265625 by ULP distance: expected a number less than or equal to 18 but got 9401
[FAIL] elu float16 1D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test elu float16 actual 0 should be close enough to expected 4.72265625 by ULP distance: expected a number less than or equal to 18 but got 17593
[FAIL] elu float16 2D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test elu float16 actual 0 should be close enough to expected 4.72265625 by ULP distance: expected a number less than or equal to 18 but got 17593
[FAIL] elu float16 3D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test elu float16 actual 0 should be close enough to expected 4.72265625 by ULP distance: expected a number less than or equal to 18 but got 17593
[FAIL] elu float16 4D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test elu float16 actual 0 should be close enough to expected 4.72265625 by ULP distance: expected a number less than or equal to 18 but got 17593
[FAIL] elu float16 4D tensor positive options.alpha
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Setting a custom alpha is not supported in tflite schema."
[FAIL] elu float16 4D tensor negative options.alpha
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Setting a custom alpha is not supported in tflite schema."
Harness: the test ran to completion.
